+++ system/modules/faq/dca/tl_faq_category.php
@@ -151,7 +151,7 @@ $GLOBALS['TL_DCA']['tl_faq_category'] = array
 			'foreignKey'              => 'tl_page.title',
 			'eval'                    => array('fieldType'=>'radio', 'tl_class'=>'clr'),
 			'sql'                     => "int(10) unsigned NOT NULL default '0'",
-			'relation'                => array('type'=>'hasOne', 'load'=>'eager')
+			'relation'                => array('type'=>'hasOne', 'load'=>'lazy')
 		),
 		'allowComments' => array
 		(