+++ system/modules/core/library/Contao/DcaExtractor.php
@@ -291,6 +291,12 @@ class DcaExtractor extends \Controller
 			$this->loadDataContainer($this->strTable);
 		}
 
+		// Return if the DC type is "File" or "Folder"
+		if ($GLOBALS['TL_DCA'][$this->strTable]['config']['dataContainer'] == 'File' || $GLOBALS['TL_DCA'][$this->strTable]['config']['dataContainer'] == 'Folder')
+		{
+			return;
+		}
+
 		$blnFromFile = false;
 		$arrRelations = array();