+++ system/modules/core/drivers/DC_Table.php
@@ -2936,6 +2936,12 @@ class DC_Table extends \DataContainer implements \listable, \editable
 				$this->Database->execute("UPDATE " . $this->strTable . " SET " . $this->strField . "=''");
 			}
 
+			// Set the correct empty value (see #6284, #6373)
+			if ($varValue === '')
+			{
+				$varValue = \Widget::getEmptyValueByFieldType($GLOBALS['TL_DCA'][$this->strTable]['fields'][$this->strField]['sql']);
+			}
+
 			$arrValues = $this->values;
 			array_unshift($arrValues, $varValue);